Updating vim config to use 256 colors (big change, really), adding plugins to git
*  This requires switching from `terminal.app` to `iterm`
*  This also requires compiling `screen` from source to add in 256 color support (see http://pjkh.com/articles/2008/07/09/osx-iterm-screen-vim-256-colors ).

Note: FuzzyFinderTextMate doesn't work for me yet, as I'm not using a version of vim that has ruby support compiled in.  I'm not sure I'm going to be _able_ to run a version with ruby support, since I suspect the machines at work will be miserable in that regard.  So, it's something I want to use, but not something I _can_ use yet.  :)